Moved python files out of `src` - Moved most of __main__ into a new module, listener.py
