+++ usb-bridge/ot3usb/__main__.py
@@ -1,114 +1,12 @@
 """Entrypoint for the USB-TCP bridge application."""
 import asyncio
 import logging
-import select
 import time
-from typing import NoReturn, Optional, List, Any
-import serial  # type: ignore[import]
+from typing import NoReturn
 
-from .src import cli, usb_config, default_config, usb_monitor, tcp_conn
+from . import cli, usb_config, default_config, usb_monitor, tcp_conn, listener
 
 LOG = logging.getLogger(__name__)
-
-# 1 second polling for select()
-POLL_TIMEOUT = 1.0
-
-
-def update_ser_handle(
-    config: usb_config.SerialGadget,
-    ser: Optional[serial.Serial],
-    connected: bool,
-    tcp: tcp_conn.TCPConnection,
-) -> Optional[serial.Serial]:
-    """Updates the serial handle for connections and disconnections.
-
-    Args:
-        config: Serial gadget configuration
-
-        ser: Handle for the serial port
-
-        connected: Whether the monitor reports the serial handle as
-        connected or not
-
-        tcp: The TCP Connection handle, which will be connected/disconnected
-        based on the serial port presence.
-    """
-    if ser and not connected:
-        LOG.debug("USB host disconnected")
-        ser = None
-        tcp.disconnect()
-    elif connected and not ser:
-        LOG.debug("New USB host connected")
-        ser = config.get_handle()
-        tcp.connect(default_config.DEFAULT_IP, default_config.DEFAULT_PORT)
-    return ser
-
-
-def check_monitor(monitor: usb_monitor.USBConnectionMonitor, msg_ready: bool) -> None:
-    """Order the monitor to update its status.
-
-    If a message is ready then the update is performed based on the message,
-    otherwise it is polled by the state file (for cases where the messages do
-    not signal before configuration completes).
-    """
-    if msg_ready:
-        monitor.read_message()
-    else:
-        # Force a poll
-        monitor.update_state()
-
-
-def listen(
-    monitor: usb_monitor.USBConnectionMonitor,
-    config: usb_config.SerialGadget,
-    ser: Optional[serial.Serial],
-    tcp: tcp_conn.TCPConnection,
-) -> Optional[serial.Serial]:
-    """Process any available incoming data.
-
-    This function will check for input from any of the input sources to the
-    USB bridge:
-        - The serial port, if it is available
-        - The UDEV message stream (usb_monitor)
-        - The TCP connection to the NGINX server, if a connection is open
-
-    Args:
-        monitor: The USB connection monitor
-
-        config: Serial gadget configuration
-
-        ser: Handle for the serial port
-
-        tcp: Handle for the socket connection to the internal server
-    """
-    rlist: List[Any] = [monitor]
-    if ser is not None:
-        rlist.append(ser)
-        rlist.append(tcp)
-
-    ready = select.select(rlist, [], [], POLL_TIMEOUT)[0]
-    if len(ready) == 0 or monitor in ready:
-        # Read a new udev messages
-        check_monitor(monitor, monitor in ready)
-        ser = update_ser_handle(config, ser, monitor.host_connected(), tcp)
-        # ALWAYS exit early if we had a change in udev messages
-        return ser
-    if ser and ser in ready:
-        # Ready serial data
-        try:
-            data = ser.read_all()
-            LOG.debug(f"Received: {data}")
-            tcp.send(data)
-        except OSError:
-            LOG.debug("Got an OSError when disconnecting")
-            monitor.update_state()
-            ser = update_ser_handle(config, ser, monitor.host_connected(), tcp)
-    if ser and tcp in ready:
-        # Ready TCP data to echo to serial
-        data = tcp.read()
-        if len(data) > 0:
-            ser.write(data)
-    return ser
 
 
 async def main() -> NoReturn:
@@ -158,10 +56,10 @@ async def main() -> NoReturn:
 
     if monitor.host_connected():
         LOG.debug("USB connected on startup")
-        ser = update_ser_handle(config, ser, True, tcp)
+        ser = listener.update_ser_handle(config, ser, True, tcp)
 
     while True:
-        ser = listen(monitor, config, ser, tcp)
+        ser = listener.listen(monitor, config, ser, tcp)
 
 
 if __name__ == "__main__":
