+++ usb-bridge/ot3usb/default_config.py
