--- usb-bridge/ot3usb/src/__init__.py
@@ -1,1 +1,0 @@
-"""Imports for ot3usb bridge."""
