+++ usb-bridge/ot3usb/usb_config.py
