+++ usb-bridge/ot3usb/usb_monitor.py
