+++ usb-bridge/tests/test_listener.py
@@ -6,8 +6,7 @@ import mock
 import select
 import serial  # type: ignore[import]
 
-import ot3usb.__main__ as ot3usb
-from ot3usb.src import usb_config, tcp_conn, usb_monitor
+from ot3usb import usb_config, tcp_conn, usb_monitor, listener
 
 FAKE_HANDLE = "Handle Placeholder"
 
@@ -33,25 +32,25 @@ def test_update_ser_handle() -> None:
     tcp = tcp_mock()
     config.get_handle.return_value = FAKE_HANDLE
     # No serial and not connection
-    assert not ot3usb.update_ser_handle(config, None, False, tcp)
+    assert not listener.update_ser_handle(config, None, False, tcp)
     tcp.disconnect.assert_not_called()
     tcp.connect.assert_not_called()
 
     # Serial but no connection
     tcp.reset_mock()
-    assert not ot3usb.update_ser_handle(config, FAKE_HANDLE, False, tcp)
+    assert not listener.update_ser_handle(config, FAKE_HANDLE, False, tcp)
     tcp.disconnect.assert_called_once()
     tcp.connect.assert_not_called()
 
     # No serial but yes connection
     tcp.reset_mock()
-    assert ot3usb.update_ser_handle(config, None, True, tcp) == FAKE_HANDLE
+    assert listener.update_ser_handle(config, None, True, tcp) == FAKE_HANDLE
     tcp.disconnect.assert_not_called()
     tcp.connect.assert_called_once()
 
     # Serial and yes connection
     tcp.reset_mock()
-    assert ot3usb.update_ser_handle(config, FAKE_HANDLE, True, tcp) == FAKE_HANDLE
+    assert listener.update_ser_handle(config, FAKE_HANDLE, True, tcp) == FAKE_HANDLE
     tcp.disconnect.assert_not_called()
     tcp.connect.assert_not_called()
 
@@ -59,18 +58,18 @@ def test_update_ser_handle() -> None:
 def test_check_monitor() -> None:
     monitor = monitor_mock()
 
-    ot3usb.check_monitor(monitor, True)
+    listener.check_monitor(monitor, True)
     monitor.read_message.assert_called_once()
     monitor.update_state.assert_not_called()
 
     monitor.reset_mock()
 
-    ot3usb.check_monitor(monitor, False)
+    listener.check_monitor(monitor, False)
     monitor.read_message.assert_not_called()
     monitor.update_state.assert_called_once()
 
 
-TIMEOUT = ot3usb.POLL_TIMEOUT
+TIMEOUT = listener.POLL_TIMEOUT
 
 SER_DATA = b"abcd"
 TCP_DATA = b"efgh"
@@ -95,7 +94,7 @@ def test_listen(monkeypatch: pytest.MonkeyPatch) -> None:
 
     # No message ready, monitor disconnected
     monitor.host_connected.return_value = False
-    assert ot3usb.listen(monitor, config, None, tcp) is None
+    assert listener.listen(monitor, config, None, tcp) is None
     monitor.update_state.assert_called_once()
     select_mock.assert_called_with([monitor], [], [], TIMEOUT)
     select_mock.reset_mock()
@@ -104,7 +103,7 @@ def test_listen(monkeypatch: pytest.MonkeyPatch) -> None:
     # Monitor has a message and is connected
     monitor.host_connected.return_value = True
     select_mock.return_value = ([monitor], None, None)
-    assert ot3usb.listen(monitor, config, None, tcp) is not None
+    assert listener.listen(monitor, config, None, tcp) is not None
     # Monitor should be manually updated
     monitor.update_state.assert_not_called()
     select_mock.assert_called_with([monitor], [], [], TIMEOUT)
@@ -117,7 +116,7 @@ def test_listen(monkeypatch: pytest.MonkeyPatch) -> None:
     select_mock.return_value = ([], None, None)
     tcp.connected.return_value = True
     monitor.host_connected.return_value = True
-    assert ot3usb.listen(monitor, config, ser, tcp) == ser
+    assert listener.listen(monitor, config, ser, tcp) == ser
     select_mock.assert_called_with([monitor, ser, tcp], [], [], TIMEOUT)
     select_mock.reset_mock()
     monitor.reset_mock()
@@ -126,7 +125,7 @@ def test_listen(monkeypatch: pytest.MonkeyPatch) -> None:
     select_mock.return_value = ([ser, tcp], None, None)
     tcp.connected.return_value = True
     monitor.host_connected.return_value = True
-    assert ot3usb.listen(monitor, config, ser, tcp) == ser
+    assert listener.listen(monitor, config, ser, tcp) == ser
     select_mock.assert_called_with([monitor, ser, tcp], [], [], TIMEOUT)
     ser.read_all.assert_called_once()
     tcp.send.assert_called_with(SER_DATA)
