+++ usb-bridge/tests/test_usb_monitor.py
@@ -5,7 +5,7 @@ import mock
 from pathlib import Path
 import os
 import pyudev  # type: ignore[import]
-from ot3usb.src import usb_monitor
+from ot3usb import usb_monitor
 
 TEST_PHY_NAME = "usbphy123"
